Clarify nuclear radius in documentation
The docs refer to "proton_rad", which is not defined anywhere. It refers to the internal value `proton_radius_ = 1.2`, but this is not accessible from the documentation. Instead, the documentation now mentions "1.2" instead of "proton_rad", clarifying the nuclear radius formula.

Furthermore, it is now documented that some specific nuclei use different default values.