Lift limitation of stochastic criterion to "elastic box"
Since the stochastic criterion is now tested for more scenarios 
including inelastic reactions 2-to-2 as well as 2-to-1 (see #274). The 
limitation to only being able to run a box with one particle species and 
a constant elastic cross section can be removed.

While there are still limits to the applicability in some cases e.g. pp 
collisions, most results obtained with the stochastic criterion lead to 
reasonable (in some cases also improved) results.