Enforce grid cell length to be larger than min. cell length
Previously, in the case of periodic boundary conditions this was not 
done, if the number of cells is increased from 1 to 2 for the peridoc 
boundaries. Since with the neighbor search all particles are anyway 
checked with each other, this did not matter for the geometric 
cirterion. But this can be relevant for the stochastic criterion, which 
does not use neighbor search, for very large cross sections.