Integrate 3-body phase space for 3-to-1 probability
Prior the phase space was approximated by a constant. The integration
was checked against the approximation that used the omega pole mass and
the case of massless incoming particles, where the integral should be
equal to M^4/2. Due to added integration, run time is slower (3x when
a lot of 3-body reactions are possible).