Scale multi-particle scatter probability with testparticles
As the proability is defined as the number of scattering over the 
product of the number of all incomping particle species, the probability 
should be scaled with 1 over the number of testpartciles to the power of 
the number of incoming particles - 1. This is because the number of 
scatterings as well as every single number of incomping particle species 
goes linearly with the number of test-particles.