Use Pythia 8 header to get a version
Let's assume that one would like to install smash into desired chroot --
/var/lib/chroot/. This chroot contains installed and supported by smash
version of the Pythia 8. However, the FindPythia cmake script would
parse the ${Pythia_prefix} to get a version and would fail to find
it. Also, this issue would arise if the Pythia 8 installed from the
package (e.g. [1]) or into /usr/local/.

This commit allows to get a Pythia 8 version even if ${Pythia_prefix}
does not contain version in it.

[1] https://packages.altlinux.org/en/sisyphus/srpms/pythia8