Implement muQ for boxes and fix tests
* Implement the charge chemical potential for boxes, it can be used for
the initialization from now on
* Clean up warnings about unused variables

Fix tests:
* grandcan_thermalizer: Increase tolerance, comment out one of the
tests, as this requires the creation of a hadron gas EoS table which now
(with the additional dimension of nQ) takes much longer than before and
is beyond the time limit of the test. The test can still be executed if
the correspondin hadron gas file is produced in advance and then only
read.
* hadrongaseos: update the functions, use all particles in the particle
list for the tests. The reason is that the solver is not able to
successfully solve the coupled Equations of the equation of state if
only the subset (pions, protons, kaons) is used. The iteration does not
converge quickly enough (if at all)