Consider form factors in photon production also if Nfrac=1
If Nfrac=1, the total cross sections are used to perform the weighting
of the produced photon process. As opposed to the differential cross
sections, those were however not corrected for the form factors. This is
now fixed.

In addition, the code was slightly resutructured and the function
'photon_cross_sections()' renamed to 'create_collision_branch()'. This
function originally only created the collision branch and computed the
photon cross section as a necessary side ingredient.

The computation of the photon cross sections is now moved to
total_cross_section().

Refs #557