Directly choose incoming postion (index) for stochastic criteron
Previously, the interaction point was randomly choosen every time, which 
lead to different interaction points on different function calls for the 
same action. Now, the index to the incoming particle, which postion is 
used, is fixed per action. 

Also, this is set earlier such that the potentials use the correct 
interaction point.