Revert some pointer iteration changes
In some cases, there are warnings when iterating over pointers like for '(const ParticleTypePtr p : mode->type().particle_types())' that p in the above example should be a pointer. However the object that is iterated over (in this case the particle_types()) already consists of pointers. So this is probably a false-positive warning. In all of these cases, the & were removed again.

Refs #581