Ensembles: adapt fermi motion and nucleons_collided logic
This fixes the logic broken by previous commit -- parallel ensembles.
The logic for 2 things is restored:
- assigning beam momenta in case of collider modus
- forbidding interactions within one nucleus

The implementation of the second was ugly anyway, so I used opportunity
to clean it. The belonging or not belonging to a nucleus is
now determined by explicit label saved in ParticleData structure.
This replaces ugly heuristics used before. If we had a clean
implementation, I wouldn't even have to modify it after
introducing ensembles.

I think that previously we didn't
want to add new variables to ParticleData structure to optimize
memory usage. That's why instead of putting projectile/target
label into ParticleData we created such ugly workaround.
Turned out to be a bad decision. Adding one byte to >50 bytes
data structure is not a dramatic increase in memory consumption.
In fact this byte may be consumed by alignment, so ParticleData
will not even become bigger.

Signed-off-by: Dmytro Oliinychenko <doliinychenko@lbl.gov>