Make grid cells slightly larger and fewer, but always equally-sized
This is necessary for the stochastic criterion that assumes the same 
volume for all grid cells. For the neigbor search in the geomtric 
criterion it is ok to have smaller cells at the grid borders. This 
change get rids of such cells and makes the (inner) grid cells larger to 
fill the grid length.