Clean up the checks and corresponding comments on grid construction
* If we would have 0 cells, which means that the system is smaller than 
min_cell_length than it is ok, to use 1 cell smaller than 
min_cell_length, since no interactions are missed with 1 cell
* An exception is the box, where we need at least 2 cells along an axis. 
We do not let them be smaller than min_cell_length.
* Because of the point above, it also does not make sense to increase 
the number of cells from 1 to 2 in the box, since this will always error 
later. Removed the corresponding if and throw.
* Removed comment about the use of half-open intervals, which no longer 
is true for equally-sized grid cells.