Block direct back-reactions for 2- and multi-body stochastic reactions
This becomes problematic for small cells, where the probability is large
and particles are therefore directly recombining. So one ends up with a
back and forth just inside a cell.

Block all 2-to-5 reaction where the 2-body initial state particles come
from same 5-to-2 reaction. Plus, block all 5-to-2 reaction where any of
the incoming particles come from the same 2-to-5 reaction.

Or commented out: Block all 2-body back-reaction and all 5-body back-
reactions where all 5 particles come from the same reaction.

To be clear: This is temporary change for debugging, backreactions
should not be blocked in general.