Only rescale the grid index factor if necessary
This avoids also problems when the grid length is zero.

Necessary here means that the index can be left too small in cases where
the grid is smaller than the minimum cell length, so for sure only one
cell along the given axis with the theoretical size of the grid is
needed. So having the index_factor be too small and always leading to
an index of 0 (so the first cell) is no problem.

This actually rolls back part of the change in commit
c652b324aaac01ec97b90c09ea4863148910c0b3 that made the index factor
to be always rescaled.