Get rid of a fishy optimization
The condition
if (smearing && sf_and_grad.first < really_small) {
  Add up contribution to density
}
was originally intended for optimization.

Indeed, premature optimization is the root of all evil. In practice, the
runtime improvement is not noticeable, but this code became a source of sneaky
bugs, because really_small is dimensionfull and has to be properly scaled with
number of testparticles and ensembles.

It seems better to get rid of this optimization completely.

Signed-off-by: Dmytro Oliinychenko <doliinychenko@lbl.gov>