Added the option to choose between the CovariantGaussian and FiniteDifference derivatives in the calculation of the DensityOnLattice currents. This was done with the help of a template specialization of the update_lattice() function.
Removed the default value for compute_gradients in update_lattice() which
had the potential to create bugs (specifically when the user does not
provide the time_step, but provides the bool for compute_gradient).

Changed the name of the DensityOnLattice variable from djmu_dx to djmu_dxmu
which is more appropriate.

Added several functions allowing for handling the current values.

Signed-off-by: Agnieszka Sorensen <awergieluk@gmail.com>