Fix the warning
In case of discrete smearing the decrease of timestep sometimes doesn't help.
The reason is the following. The relative momentum change per timestep should be
small enough for a precise solver. This is reached if the potential is smooth enough on
the spatial scale corresponding to timestep. In Gaussian smearing the smoothness
of potential is regulated by the smearing parameter, so one can improve situation
by decreasing the timestep. In discrete smearing the smoothness of potential
is regulated by lattice spacing and number of ensembles. Decrease of the timestep
is not very helpful, because relative momentum change may scale with the timestep.

Signed-off-by: Dmytro Oliinychenko <doliinychenko@lbl.gov>