Fix division by zero bug
Adding particle to Tmn with a factor was done like

1. Multiply the particle momentum by factor
2. Add pmu * pnu/p0 to Tmn

It's quite a weird way to do it, but it works. Or rather, it works unless the
factor is 0 -- then you get division by zero. Previously there was no such
problem, because there was a check in a different place that the smearing
factor is not too small. The check was a bad idea, because smearing factor has
a dimension, so checking against really_small had to be scaled with Ntest and
Nens. So we removed that check and it triggered this bug, which was there since
I designed the energymomentum class.

This refs issue #690

Signed-off-by: Dmytro Oliinychenko <oliiny@uw.edu>