Tmn: fix old todo, do not assume 0th eigenvalue is the largest
The implementation of energy-momentum tensor diagonalization relied on the idea
the Eigen sorts eigenvalues in descending order and therefore 0th eigenvalue is
the largest eigenvalue. In practice it was always fulfilled (at least to my
knowledge) and unit-tests passed.  However, recent Eigen documentation became
more detailed and it states now that eigenvalues are not ordered. To be on the safe
side I pick the largest eigenvalue explicitly now.

Thanks to Gabriele Inghirami for pointing my attention to it.

Signed-off-by: Dmytro Oliinychenko <oliiny@uw.edu>