Avoid usage of deprecated c functions
abs() -> std::abs()

This is not just some formal feature. The usage of abs() was a reason of quite
a nasty bug that resulted in suspicious proton flow v2(y) at 1 AGeV in PbPb
collisions.  Given that abs() was used in 3 different places, potentially this
commit fixes multiple very elusive bugs, which would show up when the code
is compiled by gcc, but not by clang.

This refs #705

Signed-off-by: Dmytro Oliinychenko <oliiny@uw.edu>