Add cell size specification option for stochastic criteiron
Previously the cell size could only be controlled indirectly
via the maximum cross section config option. Now, it can be
controlled directly, which is necessary for the stochastic
criterion. Results with the stochastic criterion should be
checked for dependence on the grid cell size. Only a
minimal value is given, since the length is scaled up to
fit the whole medium.

The fixed cell length is not scaled with the number of
testparticles. The default of 2.5 fm is now always used
with the stochastic criterion and matches the cell length
corresponding to the default maximum cross of 200 mb.

Also, throw error when grid is not used with the stochastic
criterion, since it should always be used with a grid i.e.
many small cells.