Do not limit cell number for stochastic criterion
For the geometric criterion it makes sense to not have less than 1 
particle in each cell, since the grid cell search is an optimization and 
the cells can be always made larger. For the stochastic collision 
criterion, the cell size is an important calculation parameter, so the 
cell size should be kept constant. The number of cells therefore cannot 
be limited as the medium grows even though this might be inefficient for 
very large systems.