Add cmake option to enable nanobenchmarking
Nanobenchmarking code is rarely used and only for very technical work
on the code. It makes then sense to only optionally include this and to
leave it out from SMASH compilation by default. Moreover, such a code
contains assembler instructions that are not supported on ARM
architectures. Hence, CMake will forbid now nanobenchmarking to be
activated on machines with such a processor.

Some documentation has been added in the CONTRIBUTING file.