Extract endianness test from source code to CMake
The PdgCode implementation needs to distinguish between little and big
endian machines and this was done testing preprocessor macros related
to the machine, which had to be put in the source code by hand by the
developer or, worse, the user. This is now extracted into the CMake,
which actually has functionality to determine the endianness of the
machine and in this way this test can be completely hidden from the
user.