Drop -mfpmath=sse flag and conditionally use -march=native
The dropped flag is most likely implied by -march=native. However, this
is not guaranteed to be supported on all machines and clang on new
Apple computers with M1 processors does not support it. Hence a test is
now performed using CMake in order to include such a flag only if the
compiler supports it. This has to be done in third party libraries as
well.