Search for new actions of outgoing particles always up to timestep end
This reverts the change in a442763, which cut the search for new actions
short and only searched until the next intermediate output. Actions of
newly produced particles in between intermediate output and timestep
end, where therefore missed if the particles were produced before the
intermediate output. This e.g. lead to increased lifetimes of resonances
in particular when running with large timesteps.