Manybody phase space sampling
Generates full n >= 3 phase space distribution. It is ready to replace the
current 3-body sampling and 5-body sampling (5-body is implemented with error,
invariant mass sampling is wrong).  It is tested: sum of generated momenta is
(sqrts, 0,0,0) as it should be, all angles are uniform in 4pi, invariant masses
are distributed according to analytical expectations.  The function is
reasonably fast, for test it took less than 10s to generate 10^6 of 5-body
samples.

Signed-off-by: Dmytro Oliinychenko <oliiny@uw.edu>