Make help local function in some test files static
Because of the -Wmissing-declarations compile flag, the compiler wants
to see declarations for all functions. In unit tests files functions
are anyhow meant to be local helpers for tests and they can (or rather
should) be static. One could fix the warning by add the prototype of
the functions before the function definition, but this is not to be
preferred in this case.