Disable variable assignments tracking in tests in debug mode
This feature has never been really used and some tests are complex and
long such that gcc reaches the limit of hash table slots and issues
annoying notes. It has been decided to disable this feature for all
tests and in case of need come back to it in the future.