Remove prevention of d' 2-to-1 back-reaction
The back-reaction was prevented beforehand, since the d' was needed for 
the cross section paramterization even when the deuteron 3-to-2 
reactions were perfomed directly via the stochastic criterion.

This was at first done to ensure consitency of the direct 3-to-2 
treatment with the treatment via the d' in multiple steps.

With the introduction of a direct dpi and dN cross section 
paramterization, it is no longer necessary to include the d' for 3-to-2 
reactions and its formation does not need to be prevented.