Move rate equations unit-test to 3rd party examples
By its purpose the rate equation solver is not a unit test.
It is a code that uses smash as a library, so it should be
in the examples folder, which is designed specifically for this.

Examples compile, but while running them, I'm getting some
crashes with invalid PdgCode. This requires further work.

Signed-off-by: Dmytro Oliinychenko <oliiny@uw.edu>