Set saturation density for nuclei correctly
Previously, the sauration density was set identically for deformed and
non-deformed nuclei. Hence, the saturation_density_ was a private member
of the nucleus class, which did not allow for it to be modified from
within a daughter class (deformed nucleus) in this case. The
saturation_density_ is now a public member and the copied functions to
set it are removed.