Put unformed particles on the grid
When employing SMASH to extract initial conditions for hydrodynamics, 
unformed particles are needed on the grid, else their 
HypersurfaceCrossingActions cannot be found. Yet, they should contribute 
to the initial hydrodynamical densities. For this a boolean is 
introduced which allows placing particles on the grid if the ICOutput is 
enabled, but else neglects them to improve performance.