Add definition of endianness macro to FindSMASH.cmake module
This is crucial to correctly compile SMASH and, in particular, if such
a macro is not defined, at the moment, the wrong code is potentially
compiled (e.g. big endian code on little endian machine). This in the
best case results in turn into an uncaught exception at run time (for
example due to an invalid PDG code).