Allow weak decays as final decays
In some cases, it would be good to allow weak decays in SMASH at the end 
of the calculation to match experimental measurements. This achieves 
this functionalty by allowing hadronically stable particles to decay 
according to all decay modes listed in decaymodes.txt.

An example of the Simga0 decay is added. The flag to decide wheter to 
perfom final weak decays is hard-coded ATM, but will be made available 
form the config.