Make endianness branching in PdgCode class more robust
The preprocessor if-clause was done on little endian architecture and
if the corresponding macro was not defined, then the else code was
executed, assuming to be on a big endian. However, a bug in CMake files
setting up the correct macro would have resulted in the else code to be
executed and possibly to wrong assumptions and run time problems.