Remove python script to fix links in CHANGELOG headers
As discussed in issue #640, doxygen at the moment does not support
links in section headers and hence in the past, when we had links in
headers of the CHANGELOG file, we needed a workaround to fix them and a
python script to do so was devised. Now we do not have links in
sections anymore and the script is not anymore needed.