Add another member to EventInfo
The reason for adding this member is to provide the information about
whether or not kinematic cuts are imposed for SMASH initial conditions
to the output functions. Like this the warning about particles that have
not crossed the hypersurface at the end of the evolution can be
surpressed since it is expected that some will not pass it (and thus
remain in the system) if kinematic cuts are imposed.