Fix bug: "smash --version" not always providing a version
The version information printed to screen by the --version option of
the smash executable was relying on VCS (git) information, but SMASH
might be used after having downloaded the source without git
information. In this case the --version option was not providing a
version and this was invalidating the hashing mechanism of tabulations
in the main function.

Now the CMake project is declared using the VERSION option. This was
introduced in CMake 3.0 and this implies that CMake is now issuing a
warning for third-party libraries that request older CMake versions as
minimum version (Einhard and YAML). It is fine to accept this, since it
is planned to update YAML soon to version 0.7, which will make the
warning disappear. Einhard can be patched, because it will hardly be
updated.