Do not always set and use GIT_BRANCH macro
Since SMASH might be used without VCS (git) information available, it
does not make always sense to define a GIT_BRANCH macro. This is
achieved by setting it as CMake variable to either the branch name or
to NOTFOUND and let then CMake conditionally define it or not  as C++
macro at configure_file time into config.h.in via the #cmakedefine
mechanism.