Add option Minimum_Nonempty_Ensembles (#840)
This option allowes the user to specify the number of ensembles
in which at least one collision took place rather than defining
the number of events. As an additional parameter the maximum
number of ensembles to be calculated need to be given. This
is to prevent smash from running infinitely.