Improve requirement of C++ standard in CMake setup
An intricate compilation test was done before and this was not
catching exclusively an error about the C++ standard. SMASH contains
some advanced functionality about compiler flags in the cmake folder
and these are now used to test if the used compiler support the
standard. This check is not needed, but it is user friendly to have
it and fail early instead of failing at compilation.