Rewrite CMake code about compiler flags in various build modes
The way compiler flags were added in the offered build configurations
was against the CMake philosophy. On one side CMake provides by default
a given set of configuration modes with sensible default flags, and on
the other side we were force-caching variable contentes and so only
apparently giving the user the possibility to change them. This was
potentially misleading and has been changed. Actually, now only flags
for the profiling build configuration are explicitly set. Finally, all
compiler flags that are added are tested and skipped if not supported
warning the user.