Impose position independent code using CMake functionality
So far this was imposed hard-coded as compiler flag (-fPIC) everywhere
but this was not portable and easy to break. CMake offers a way to
request position independent code and this was used removing the
compiler flag wherever given. Please, note that external dependencies
like Pythia should also compiled in an appropriate way and, hence, the
compiler flag -fPIC was left on that side (e.g. in the README),
although not universal in terms of compilers.