Extract test about formatting in public action
The test about the correct formatting of the codebase in the public
repository can be done only once and not for all combinations of OS and
compilers. We now restrict the setup of the python environment to
Ubuntu 20.04 and we install only there both clang- and cmake-format
running then the formatting checks.