Require gcc-5 at least and remove action test with gcc-4.8
Since we need to do a bug-fix release, we take this opportunity to
require a slightly more recent compiler, so that the user can build the
full codebase, including tests. The GitHub action on Ubuntu-18.04 with
gcc-4.8 has been removed since not needed any more (in any case it was
failing at compilation because of the unit testing framework correctly
using lambda functions in variadic templates in a not-yet-supported way
by gcc-4.8).