Update copyright years fixing them by using git log history
Many copyright statements contained wrong dates probably because of
copy and paste at creation time. It has been now decided to include in
all files only the modification years that have been extracted from the
repository history using the --follow option of git-log in order to
track as much as possible past renames of the files.