Increase required C++ standard to C++17 and adjust README
Since in the top-level CMake file we use the CMAKE_CXX_STANDARD to
impose the given standard on all targets, this would impose it also for
third-party libraries and it would do so silently. Therefore, such a
variable is temporally unset and then restored in the CMakeLists.txt
file in the 3rdparty folder.