Add missing header file for compilation on Ubuntu
Although compilation on Apple machines was fine, some files were
failing to compile on FIAS machines, due to std::filesystem::path not
being found. Including the filesystem library in a couple of places
fixed the problem. Please, note that with boost the path object was
forward declared while the std class is not and this is probably the
reason why the simple replacement of the include directive was not
enough. This is OS dependent, though.