Check for minimum required version of supported compilers
Previously this was not done and only a check about whether the
compiler supported the needed C++ standard was done. However, such a
check was way not optimal, because it was just checking if a trivial
empty main function was compiling with the -std=c++17 flag. This might
have failed if the user passed an invalid compiler flag e.g. via CMake
variables on the command line, resulting in a misleading error.
Furthermore, now that we use C++17 standard, we do not require the bare
minimum compiler version for such a standard, but a more recent one,
since we need the filesystem library that was introduced later by
compilers. Therefore it makes sense to implement a check on the
compiler version.