Avoid using -Wduplicated-branches compiling lattice test
This flag that was introduced in GNU version 7 is by default used with
GNU compiler, but it turned out to make compilation hang for such a
test. Since the code that makes the compiler hang is the last call to
iterate_sublattice() function in the TEST(iterate_in_rectangle), a very
naive guess is that the compiler is not able there to handle the
if-else clause in the lambda function passed to iterate_sublattice().
This might well be the case, because the branches there are made of
complex call to the COMPARE macro of the unit tests framework.