Bump minimum required CMake version to 3.16
In the 3rdparty CMakeLists.txt file, I previously made use of the
CMAKE_MESSAGE_INDENT macro, which has been introduced in CMake 3.16 and
I did not bump at the same time version. So e.g. for version 3.14 the
user would get an error message. It has been decided to adjust the
requirement to have for future work a pretty recent CMake version which
should not be changed soon after next release.