Remove not anymore needed std::move on return type
As a comment in the code was stressing, this was needed for old
compiler versions like GNU v4.8. Since now we upgraded to C++17 and the
minimum compiler versions have increased, it is time to get rid of this
superfluous std::move calls. For the records, I report here for the
records some research that basically explains why with the C++11
standard those std::move calls were needed.

The C++ standard rule to be applied here is 12.8/32, which roughly
speaking says that, when returning an object, then under certain
circumstances, the returned object is first tried to be treated as an
rvalue and, only if that fails, it is then treated as the lvalue it
is. Now, these circumstances in C++11 were that the returned object and
the return type to be the same cvunqualified type. This has been
relaxed in C++14 where in addition the same rule also applies if the
returned object is a local one.
In the two functions considered here the returned pointer is not of the
same cvunqualified type of the return type and this violates the
circumstances in C++11. Therefore, in C++11 the code without std::move
is not compiling since no std::unique_ptr constructor would be
viable. However, the return object is a local type and this means that
from C++14 the return type is treated as an lvalue first, letting
overload resolution (to decide which constructor to call at return
time) find a viable candidate to call, then, even without std::move.