Require version 13 of clang-format and clean its setup file
The .clang-format file was plenty with commented out options that were
not used and hence the value of Google style was used unchanged. Since
such a code was cluttering the file content, I removed it. Only for
those options that are changed w.r.t. to Google default, I left some
explanation, which is taken from clang-format documentation.