Remove path to include folder into include directives
It is unnecessary and actually against Google style guide to give the
full relative path to the include folder from the present working
directory in the include directive. This is also potentially requiring
a change in case the file is moved to a different folder. The compiler
will be able to find the header files, since the include folder is
passed to it using the -I option.