Remove unnecessary YAML library files
The test and util YAML folders have been deleted, because we
explicitly set up YAML not to use them. Hence, in YAML CMake setup,
these folders do not get loaded, which in turn means that those folders
might also not exist without causing any problem. This change will make
future releases thinner and a note in the 3rdparty maintenance README
has been added to be sure to keep such a change in future. As side
note, unfortunately, such a change has not been done straight away in the
past and all the unneeded files got into the repository history.