Add proof of concept of new Key and InputKeys classes
The latter is meant to contain static members only and it will be the
database of ever existed SMASH input keys (including removed and
deprecated ones). The rationale behind this choice is that the Key
object contains metadata about when the key has been deprecated or
removed, if any. Everything is traced in terms of released SMASH
version numbers.