Implement proof of concept for configuration validation
The validation function should take a reference to a const
Configuration or be a const method of the Configuration class. However,
at the moment the class is still missing some functionality to read
sections without extracting them. Hence here a first implementation is
done, which however does not work on a constant object and wrongly
modifies it. Furthermore the extract_sub_configuration is also called
to get keys with a single vale and not only sections; hence tests are
not all passing, since in the validation an exception is thrown.