Refactor extraction of Rivet parameters from SMASH configuration
The OutputParameters object was containing a piece of the SMASH
configuration and this was parsed one level down, in the RivetOutput
object. I think there is no reason to entangle the RivetOutput object
with the Configuration functionality and the OutputParameters should
simply contain parameters, as the name says. I then introduced a POD
struct for Rivet parameters and handed this over to the RivetOutput
constructor. The SMASH Configuration parsing is then done in the
OutputParameters object at construction.