Fix particle and decay string in tabulation hash
The particle and decay string were always empty as they were taken after 
the take from the configuration. Now they are defined beforehand. 

The consequences of this bug are minor: It could happen in some cases, 
if one changes decaymodes and partices that the tabulation were not 
recalculated and cached leading to some integrals missing, if particles 
and decays were added in between. They then were calculated on-the-fly 
during the actual runtime, leading to (I suspect only very slightly) 
longer runtime.