Let Configuration::take method remove empty sections
This change is intended to implement a naturally expected behavior.
Although it might be argued that this is a pessimization of the code
since now the method has to check and clean at every call, this is
hardly to be in a critical place, especially since take is called once
at most per parameter. It can always be reversed in the future. The
method has not been renamed with a more descriptive name because the
behavior is still part of the "taking" action (and such a method is
largely used in the codebase).