Fixes energy violation issue in IC output
Sometimes in processes managed by Pythia the energy conservation can be violated. This creates a problem in the final check of the Initial_Conditions output, that compares the total energy of the initial particles with the total energy of the particle which have crosses the isochronous hypersurface. In this commit we introduce an additional member in the Experiment class, total_energy_violated_by_Pythia_, which keeps track of the amount of energy density violation, taking it into account when performing the aforementioned final check.