Fixes conserv. law issue in crossing hypersurface
When using the Initial_Conditions output in some relatively rare cases the particle crossing an isochronous hypersurface can have an invariant small so small that its momentum is considered conserved even if actually the particle has been removed from the system. However, since the check that the particle has been removed is based on the non conservation of momentum and quantum numbers, this creates a problem if the particle has the various quantum numbers, like charge, baryon number and isospin3, all 0, because, if everything appears conserved, SMASH concludes that actually the particle has not been removed and terminates with an error. In this commit we remove one of the two checks on the four momentum conservation, which ultimately uses the function almost_equal_physics, defined in src/include/smash/numerics.h, and slightly modify the other.