The Configuration validation is now done as internal const method
Making the validate funtion a public member of the Configuration class
allows lots of simplification, since it is possible to take advantage
of YAML library functionality. In this way the recursive implementation
to get the list of labels per key becomes much simpler and it is
possible to have the validation as constant method, without having to
add any functionality to the Configuration class.