Declare keys of General section as existing and move documentation
This is the first commit of a series that will move documentation about
valid keys in the input configuration file from around in the codebase
to a single place. Doing this refactoring, InputKeys static members
will be created and the validation mechanism will slowly be
completed. Note that the documentation is also reviewed and improved,
trying to make it consistent in layout. For the moment, version 1.0 is
used as version in which keys have been introduced and a decision about
whether to try to find out when each key was introduced will be taken
later.