Fix logic checks in deformed nuclei creation
Projectile and/or target nuclei deformation in collider mode can be
requested as automatic or manual and, depending on this, other
parameters must or must not be specified. Before, in case of manual
deformation, all parameters had to be specified. However, since some
have been added, it has been now decided to only request that at least
one beta parameter should be given in the input file. If, instead, the
deformation is requested to be automatic, then no parameter should be
specified.