Fix validation problem due to keys taken as maps
Technically speaking, a section in a YAML file is a key whose value is
a map of key-values and, in turn, the section key values can be as well
sub-sections, and so on. In the validation algorithm, the configuration
file is processed in order to build a list of key labels, namely all
labels from the top-level of the YAML file that identify a value. In
this process, maps are automatically identified as sections and
iterated over recursively. Therefore, if a SMASH key value is meant to
be a map, this map will be unrolled in the validation algorithm and it
has to be adjusted afterwards. To do so, it is enough to look in the
database of input keys for those that have a std::map as type and use
their labels to adjust the "unrolled configuration list".

Because of the InputKeys::list type, the needed code is not the most
beautiful ever, but amplification comments have been added where
needed. A KeyLabels alias of std::vector<std::string> has been
introdueced as well to make the code more expressive.

Now all validation tests pass.