Avoid intermediate output only after configured end time
We only want to produce an intermediate ouput before the end time as the 
same output is produced by the final output as well. With the change of 
the run_time_evolution to take an end time as an argument (t_end), it 
might be that this is not the configured end time (at which a final 
ouput is produced) as SMASH is e.g. run in timesteps externally. 
Therefore intermediate output was avoided even though we do not print a 
final outout. With using the configured end time (end_time_) at which 
final output is produced, this swalloing of intermediate output is 
avoided.