Avoid usage of std::transform_reduce algorithm
I had not noticed that std::transform_reduce is intrinsically a C++17
parallel executed algorithm and it would require more recent compilers
version, e.g. GNU version 11, and we do not want to raise the
requirement so much for so little (the place where this is used will
never be a bottleneck in performance). Hence the code has been
refactored using a raw loop instead of the algorithm.