Replace GitHub emoji name with unicode symbol
This is again a kind of patch to make the emoji be displayed both in
GitHub and in Doxygen output. In Doxygen the emoji should be specified
using the \emoji command followed by the emoji name, but this would
lead to have the \emoji string when the file is rendered on GitHub. The
patch seems to work, bot sure for how long, though.