Add optional environment variable to improve FindSMASH module
The SMASH_BUILD_DIR environment variable can now be used to tell CMake
where to look for some files. If the environment variable is not set,
it is assumed to be "${SMASH_DIR}/build" (as it was before). A check on
SMASH_DIR to make the script fail in a more user-friendly way if it is
not defined has been added.